Split api_converters.hpp
